moving coordinator responsibilities to master
