+++ ipc/messages.h
@@ -19,8 +19,8 @@ static int MSQ_PERMISSIONS = 0666;
 static long mtype_put = 1;
 static long mtype_point_query = 2;
 static long mtype_range_query = 3;
-static long mtype_kill_master = 98;
-static long mtype_master_dying = 99;
+// static long mtype_kill_master = 98;
+// static long mtype_master_dying = 99;
 
 typedef struct assigned_vector {
     vec_id_t vec_id;
@@ -38,10 +38,6 @@ typedef struct msgbuf {
     assigned_vector vector;
     vec_id_t point_vec_id;       /* for point query */
     range_query_contents range_query; /* for range query */
-    //vec_id_t range_lower[1024];
-    // unsigned int range_upper[128];
-    // char ops[128];
-    // unsigned int num_ranges;
 } msgbuf;
